Create practice makes perfect.rb
Define a greeter method that takes a single string parameter, name, and returns a string greeting that person.

Make sure to use return and don’t use print or puts.

2.
Define a by_three? method that takes a single integer parameter, number, and returns true if that number is evenly divisible by three and false if not.