Create sorting.rb
Inside your method, add a line that calls .sort! on the arr array. Since it is the last line of the method, the sorted array will be returned.

Now test your work! After your method, make a new array called numbers that contains a few numbers like the example above.

Call your alphabetize method with your numbers array as the parameter and puts the result to the console.